Lots of small updates to outcomes makefile and scripts, mainly to provide unique filenames for lineplots and map outputs. Also tracked some small spatial files in persistent_data/spatial_data

+++ minos/modules/S7MentalHealth.py
@@ -100,7 +100,7 @@ class S7MentalHealth(Base):
 
         # Draw individuals next states randomly from this distribution.
         # Update population with new income
-        self.population_view.update(men_health_prob_df['S7_mental_health'].astype(int))
+        self.population_view.update(men_health_prob_df['S7_mental_health'])
 
     def calculate_S7_mental_health(self, pop):
         """Calculate income transition distribution based on provided people/indices
